feat: implement testing by matching substring
