+++ src/main.rs
@@ -7,6 +7,8 @@ mod translate;
 mod utils;
 mod optimizer;
 
+mod testing;
+
 pub mod built_info {
     include!(concat!(env!("OUT_DIR"), "/built.rs"));
 }
@@ -24,7 +26,6 @@ use std::io::{prelude::*, stdin};
 use std::path::PathBuf;
 use std::process::Command;
 use std::{fs, io};
-use rayon::prelude::*;
 
 fn get_version() -> &'static str {
     built_info::GIT_VERSION.unwrap_or(built_info::PKG_VERSION)
@@ -139,21 +140,21 @@ struct DocsCommand {
 }
 
 #[derive(Args, Clone, Debug)]
-struct TestCommand {
+pub struct TestCommand {
     /// Input filename or directory ('-' to read from stdin)
     #[arg(default_value = ".")]
-    input: PathBuf,
+    pub input: PathBuf,
 
     /// Arguments passed to Amber script
     #[arg(trailing_var_arg = true)]
-    args: Vec<String>,
+    pub args: Vec<String>,
 
     /// Disable a postprocessor
     /// Available postprocessors: 'bshchk'
     /// To select multiple, pass multiple times with different values
     /// Argument also supports a wildcard match, like "*" or "b*chk"
     #[arg(long, verbatim_doc_comment)]
-    no_proc: Vec<String>,
+    pub no_proc: Vec<String>,
 }
 
 fn main() -> Result<(), Box<dyn Error>> {
@@ -186,7 +187,7 @@ fn main() -> Result<(), Box<dyn Error>> {
                 handle_completion();
                 0
             }
-            CommandKind::Test(command) => handle_test(command)?,
+            CommandKind::Test(command) => testing::handle_test(command)?,
         }
     } else if let Some(input) = cli.input {
         let options = CompilerOptions::from_args(&cli.no_proc, false, false, None);
@@ -301,159 +302,6 @@ fn handle_docs(command: DocsCommand) -> Result<(), Box<dyn Error>> {
     }
 }
 
-fn handle_test(command: TestCommand) -> Result<i32, Box<dyn Error>> {
-    let input_path = command.input;
-    let mut files = vec![];
-    if input_path.is_dir() {
-        find_amber_files(&input_path, &mut files)?;
-    } else {
-        files.push(input_path);
-    }
-    files.sort();
-
-    // Discovery phase
-    let mut tests = vec![];
-    for file in &files {
-        let code = match fs::read_to_string(file) {
-            Ok(c) => c,
-            Err(e) => {
-                println!("Failed to read file {}: {}", file.display(), e);
-                return Ok(1);
-            }
-        };
-
-        let options = CompilerOptions::from_args(&command.no_proc, false, true, None);
-        let compiler = AmberCompiler::new(code.clone(), Some(file.to_string_lossy().to_string()), options);
-        
-        match compiler.tokenize() {
-            Ok(tokens) => {
-                match compiler.parse(tokens) {
-                    Ok((_, meta)) => {
-                        for name in meta.test_names {
-                            tests.push((file.clone(), name, code.clone()));
-                        }
-                    }
-                    Err(e) => {
-                        e.show();
-                        return Ok(1);
-                    }
-                }
-            }
-            Err(e) => {
-                e.show();
-                return Ok(1);
-            }
-        }
-    }
-
-    let total = tests.len();
-    if total == 0 {
-        println!("No tests found");
-        return Ok(0);
-    }
-
-    let failed = std::sync::Mutex::new(vec![]);
-
-    tests.par_iter().enumerate().for_each(|(i, (file, name, code))| {
-        let test_name_display = if name.is_empty() {
-            format!("{}", file.display())
-        } else {
-            format!("{} ({})", file.display(), name)
-        };
-
-        let options = CompilerOptions::from_args(&command.no_proc, false, true, Some(name.clone()));
-        let compiler = AmberCompiler::new(code.clone(), Some(file.to_string_lossy().to_string()), options);
-
-        let result = match compiler.compile() {
-            Ok((_, bash_code)) => {
-                match Command::new("bash").arg("-c").arg(&bash_code).output() {
-                    Ok(output) => {
-                        if output.status.success() {
-                            Ok(())
-                        } else {
-                            let err_msg = format!(
-                                "{}\n{}",
-                                String::from_utf8_lossy(&output.stdout),
-                                String::from_utf8_lossy(&output.stderr)
-                            ).trim().to_string();
-                            if err_msg.is_empty() {
-                                Err(Message::new_err_msg("(No output)".dimmed().to_string()))
-                            } else {
-                                Err(Message::new_err_msg(err_msg))
-                            }
-                        }
-                    }
-                    Err(e) => Err(Message::new_err_msg(format!("Error executing bash: {}", e))),
-                }
-            }
-            Err(e) => Err(e),
-        };
-
-        match result {
-            Ok(_) => {
-                println!("[{}/{}] {} {} ... {}", i + 1, total, "✓".green(), test_name_display, "Success".green());
-            }
-            Err(msg) => {
-                println!("[{}/{}] {} {} ... {}", i + 1, total, "×".red(), test_name_display, "Failed".red());
-                failed.lock().unwrap().push((i + 1, test_name_display, msg));
-            }
-        }
-    });
-
-    let failed_vec = failed.lock().unwrap();
-    if !failed_vec.is_empty() {
-        println!();
-        for (i, name, msg) in failed_vec.iter() {
-            println!("[{i}] {} failed with:", name);
-            if let Some(m) = &msg.message {
-                println!("{}", m);
-            } else {
-                msg.show();
-            }
-            println!();
-        }
-
-        println!("{}", "Summary of failed tests:".red());
-        for (_, name, _) in failed_vec.iter() {
-            println!("{} {}", "×".red(), name);
-        }
-        println!();
-    }
-
-    let passed_count = total - failed_vec.len();
-    let failed_count = failed_vec.len();
-
-    if failed_count == 0 {
-        println!("{}", " Success ".white().on_green());
-    } else {
-        println!("{}", " Failure ".white().on_red());
-    }
-    println!("{} tests passed | {} tests failed", passed_count, failed_count);
-
-    if failed_count > 0 {
-        Ok(1)
-    } else {
-        Ok(0)
-    }
-}
-
-fn find_amber_files(dir: &PathBuf, files: &mut Vec<PathBuf>) -> std::io::Result<()> {
-    if dir.is_dir() {
-        for entry in fs::read_dir(dir)? {
-            let entry = entry?;
-            let path = entry.path();
-            if path.is_dir() {
-                find_amber_files(&path, files)?;
-            } else if let Some(ext) = path.extension() {
-                if ext == "ab" {
-                    files.push(path);
-                }
-            }
-        }
-    }
-    Ok(())
-}
-
 fn handle_completion() {
     let mut command = Cli::command();
     let name = command.get_name().to_string();
@@ -480,7 +328,7 @@ fn handle_err(err: std::io::Error) -> ! {
 
 #[inline]
 #[allow(unused_must_use)]
-fn render_dash() {
+pub fn render_dash() {
     let str = "%.s─".dimmed();
     Command::new("bash")
         .arg("-c")
